+++ src/helpers/code-lenses.helpers.ts
@@ -1,5 +1,6 @@
 export const annotations = ["only", "skip", "fixme"];
 export const expectOptions = ["soft"];
+export const describeOptions = ["serial"];
 
 export function regexpIsTest(annotations: string[]): RegExp {
   return new RegExp(`^\\s*(it|test)(?:\\.(?:${annotations.join("|")}))?\\s*\\(\\s*[\\r\\n]*\\s*['"\`]`, "m");
@@ -7,7 +8,9 @@ export function regexpIsTest(annotations: string[]): RegExp {
 
 export function regexpIsSuite(annotations: string[]): RegExp {
   return new RegExp(
-    `^\\s*(describe|test\\.describe)(?:\\.(?:${annotations.join("|")}))?\\s*\\(\\s*[\\s\\S]*?['"\`]`,
+    `^\\s*(describe|test\\.describe|describe\\.serial|test\\.describe\\.serial)(?:\\.(?:${annotations.join(
+      "|"
+    )}))?\\s*\\(\\s*[\\s\\S]*?['"\`]`,
     "m"
   );
 }