+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "playwright-helpers" extension will be documented in 
 
 ## [1.2.11] - 2024-10-XX
 
+- feat: code lense for serial mode
 - fix: display name of trace.zip in trace view
 
 ## [1.2.10] - 2024-10-02